dataplatform/projects/reusables/python/requirements.txt
redis>=5.0.0
google-genai>=0.2.0

